src/edge/EdgeSolution/modules/webmodule/backend/requirements/base.txt
-r ./core.txt
asgiref==3.2.10
django==3.0.8
djangorestframework==3.11.0
drf-yasg2
drf-url-filters==0.5.1
pillow
azure-cognitiveservices-vision-customvision==3.0.0
requests
django-cors-headers
azure-iot-device~=2.0.0
python-dateutil
opencensus
opencensus-context
opencensus-ext-django==0.7.2
opencensus-ext-stackdriver==0.7.2
channels==2.4.0
django-extra-fields==3.0.0
django-extensions==3.0.9
kubernetes
azure-iot-hub
azure-identity
azure-mgmt-resource
opencensus-ext-azure==1.1.5
azure-storage-blob
azure-eventhub